actor/rpi/requirements.txt
paho
RPi.GPIO
